pyproject.toml
[project]
name = "chatbot"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.34.0",
    "openai>=1.12.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
]
